Update space-place-gazetteers.md
Paragraphs 21-29. These instructions read as if the reader is meant to fill in information as well as the field names, but the following section suggests these should be done separately (as does the first screenshot of the columns empty except for the headers). I suggest removing the instructions on the specific data entry and incorporating them from paragraph 34 onwards.
	SG – I’ve made updates as suggested, splitting the steps between the table header stage and the filling in the tables stages

Paragraph 32: including this as an alert box would keep the flow of the lesson.
	SG - done

Paragraph 33: Use a separate 'data dictionary' header.
	SG - another header here would be awkward so I've also put this in an alert box

Paragraph 35: replace screenshot with markdown table
	SG - done

Paragraph 34: Start a new section here to distinguish populating the spreadsheet with data from building the spreadsheet field names (see above).
	SG - done

Paragraph 36: remove the period after before To on the final line of the paragraph.
	SG - Done

Paragraph 38: replace screenshot with markdown table
	SG -Done

Paragraph 40: explain ISO code and where they can be found.
	SG - done
Paragraph 41: replace screenshot with markdown table.
	SG - done

Paragraph 43: replace screenshot with url to https://whgazetteer.org/search/
	SG - done

Paragraph 44: specify that the reader should search for Tudela in the search box.
	SG - done

Paragraph 45: replace the hand-drawn arrow with a shape/clipart
	SG - done

Paragraph 52: I was slightly confused until I remembered that Tudela is both the author and the place! Might be worth specifying his full name here.
	SG - done

Paragraph 53: In contrast to the rest of the lesson, there is a lot of dense information and jargon in this paragraph. I suggest rewriting, explaining each term and method in a little more detail.
	SG - done

NOTE - all figure captions renumbered by SG, new image assets will be sent by email